ProductProject/requirements/base.txt
django==4.2.1
Pillow==9.5.0
python-decouple==3.8
djangorestframework==3.14.0
requests==2.31.0